requirements.txt
# FastAPI and Web
fastapi>=0.109.0
uvicorn>=0.27.0
jinja2>=3.1.0
python-multipart>=0.0.6

# AI and Embeddings
openai>=1.12.0
tiktoken>=0.5.2

# Vector Database
pinecone>=5.0.0

# Web Search
tavily-python>=0.3.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
pgvector>=0.2.4
alembic>=1.13.0

# Environment
python-dotenv>=1.0.0

# Markdown rendering
markdown>=3.5.0

# Headless Browser (for JS-rendered pages)
playwright>=1.40.0

# PDF Parsing
pypdf>=3.17.0

# Utilities
httpx>=0.26.0
aiofiles>=23.2.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
nltk>=3.8
